initial structure including rl for sim param space
